heroku: bash: node: command not found #2
According to this:

   http://stackoverflow.com/questions/39237596/deploying-app-on-heroku-node-command-not-found

Use the below command to set the application of type nodejs.

   heroku buildpacks:set heroku/nodejs

When I run this from the command-line I get:

   Buildpack set. Next release on limitless-tundra-27906 will use heroku/nodejs.
   Run git push heroku master to create a new release using this buildpack.

However, my 'git push heroku master' will see no file deltas, so I'm adding
some white space to a file.  Was hoping that adding the build-pack fu to the app.json
was sufficient, but apparently not.