update utility for pushing endpoint data to a mysql bulk-loadable file
We have this nice utility:

   normalizeData2File.js

which was reading ajax requesting data from an endpoint and writing the
results to a file with the intention of using that file to bulk-load
a relational database with normalized data.

However the stringified json in the output file was not well formed.
With this commit, the output is pleasing to mysql's 'bulk load from json'
feature.